Change "Oxidation states" to "Commonly occurring oxidation states" (#69)
The data do not exhaustively list all possible oxidation states. E.g. for Iron, Wikipedia lists states −4, −2, −1, 0, +1, +2, +3, +4, +5, +6, +7, and Britannica +2, +3, +4, +6. The Mendeleev db list only +2, +3 , which I personally find reasonable - but that should be reflected in the name of the field.

The data may still need some attention. If only oxidation states of an element in chemical compounds are taken into account, then for inert gases no zero oxidation states should listed. Otherwise, for every element a zero oxidation state must be listed. For Xe, only zero oxidation is listed, although stable Xe compounds are known (and used practically). For Cr, I'd suggest adding +4 to Cr's oxistates, as CrO2 is a common compound (used for magnetic tapes).

I could look over the data, however I'm not a chemist, and a decision which oxistates can be considered as "common" is a subjective one.